engine: Modified network static IP validation
It used to not allow changing the management network's IP address
whenever the hostname was given as an IP address, but actually
problems only arise if the old management network's IP address was
equal to the hostname. However, this validation should be applied to
ANY network, not just the management network.

Also modified the tests accordingly, as well as fixed some logic that
was supposed to give existing NICs different IP addresses but I don't
think did anything.

Change-Id: I85f873fe4e87d16ec60da5bd8f9d6c90ee1e0030
Bug-Url: https://bugzilla.redhat.com/989360
Signed-off-by: Lior Vernia <lvernia@redhat.com>