webadmin: Rendered checkbox headers resizeable
Previously, CheckboxHeader wasn't resizeable. Created a class
ResizeableCheckboxHeader to fix this, a wrapper for CheckboxHeader
that inherits from ResizableHeader (classic case for multiple
inheritance, but used composition due to Java limitations).

A CheckboxHeader is automatically wrapped with a
ResizeableCheckboxHeader when it is added to a ColumnResizeCellTable
that has column resizing enabled, as in such a case it would be
expected that the checkbox headers are resizeable as well.

Change-Id: Ie442350893282e7838e5216d277718e654f1d884
Bug-Url: https://bugzilla.redhat.com/1009745
Signed-off-by: Lior Vernia <lvernia@redhat.com>