core:  in PM only selecting apc or ipmilan...
in PM only selecting apc or ipmilan changes the options, choosing other
types is static

The reason for that was that fence options mapping were called with
'general' version and since this key is versioned, the static value for
this key in ConfigValues.java was returned. This value misses some of
the recent agents added in 3.2 and 3.3 and cause this BZ.

This patch fix this issue by getting 6the fence options mapping from the
database for the given cluster version.

Change-Id: I2f815e90d95e8cd21aca51a803dccde8079dee5d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1014513
Signed-off-by: Eli Mesika <emesika@redhat.com>