restapi: Don't add permissions via user/roles
Currently permissions can be added using the /api/users/{user:id}/roles
collection and doing it results in adding system permissions. The
support for adding system permissions has been moved to /api/permissions
in a previous patch. This patch makes the /api/users/{user:id}/roles
read only so that the only way to add system permissions will be the
/api/permissions resource.

Change-Id: Id0acfd16d14965cfb7988556ed1b0ec1d3605c1d
Bug-Url: https://bugzilla.redhat.com/1018552
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>