core: deadlock and related changes in domain monitoring/InitVdsOnUp
The following patches fixes few issues related to the given bug:

1. Deadlock between domain failover process and InitVdsOnUp flows -

Thread 1 (domain failover process) - runs within the event queue (a),
attempts to call SetVdsStatus command that waits for lock on
VdsManager.getLockObj() (b)

Thread 2 - InitVdsOnUp - runs through the VdsManager.OnTimer with lock on
VdsManager.getLockObj() (b) and waits for event submitted to the event
queue to be executed (a)

The implemented solution is to call SetVdsStatus executed from thread 1
on a new thread which will also reduce the wait time for the mutex.

2. Domain monitoring information should be processed from hosts that are
 "Up", if the host is not UP it means that something is wrong with
that domain regardless (for example, if the host was moved to non-op
status because of problematic report of domain, there's no reason to
analyze each time its domain report).

3. Reduce calls to SetVdsStatus thus reducing acquired locks and wait
time in case that the host isn't UP (if the host is already not up
because
of some other reason, there's no need to change it's status).

Change-Id: I334b450fcf29ed57675ec1f2e4bf3bc43f31b4aa
Bug-Url: https://bugzilla.redhat.com/965972
Signed-off-by: Liron Aravot <laravot@redhat.com>