webadmin: Remove vNIC items changed listener before flush
Upon flushing the AddRemoveRowWidget, it sets the backing ListModel's
items according to the entries in the widget. However, in the add/edit
VM dialog, since it is listening on the items changed event, this
re-initializes the widget just before it closes, thus adding a new
vNIC row. This patch moves the listening logic into
ProfilesInstanceTypeEditor and removes the listener prior to flushing
the widget, in order to avoid that flicker.

Change-Id: I5f153a3b19e467d41131015c7e2819c3ce365c22
Bug-Url: https://bugzilla.redhat.com/1020755
Signed-off-by: Lior Vernia <lvernia@redhat.com>