webadmin: Expose read-only disk functionality in UI.
Allow user to configure a read only property of a disk in the scope of
VM. User can configure this property in Add/Edit VM disk popup windows
and see it's value in related column under VM Disks subtab.

Note that shared disk can be read only in scope of one VM and read
write in scope of another therefore this property is meaningful in VM
scope only. All other appearances of a disk in UI won't provide neither
option to configure read only property nor visualize it's value.

Change-Id: I2c8d98df5e9b7aa8130b45b1ca036219175b70bf
Bug-Url: https://bugzilla.redhat.com/878705
Feature-Page-Url: http://www.ovirt.org/Features/Read_Only_Disk
Signed-off-by: Vered Volansky <vvolansk@redhat.com>
Signed-off-by: Sergey Gotliv <sgotliv@redhat.com>