core: export memory snapshot volumes for diskless vms
When exporting a diskless vm, the command is ended right away if the vm
doesn't have any disks for export, as no tasks would be created.
The VM might have no disks, but have snapshots that contain memory
snapshots, those memory volumes need to be exported as well otherwise the
export data would be partial only.

Change-Id: Ie8afa364925dc8a8cdfc1f64939a6f28945fe043
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1029045
Signed-off-by: Liron Aravot <laravot@redhat.com>