core: empty quota causes FK violation
When creating a disk in Foreman, the disk object is sent via
REST with empty string in id attribute:
<quota id=""/>

REST creates an empty UUID (000...) out of empty string, and causes
FK violation, since there is no quota with that id.

This patch fixes it by ignoring empty quota for adding disk.

Change-Id: Ic488486edc7c5602d24e804c9a89aeccc98fc32b
Bug-Url: https://bugzilla.redhat.com/1022942
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>