core: quota permissions isn't enforced in add VM
When creating a VM or one of the command's derivatives (add VM from .., etc.),
a consume quota check should be performed on the designated quota,
Currently when creating a VM, there isn't any permission check on that quota, the
check if performed only if the quota is consumed (i.e running a vm).

Also when running a VM there shouldn't be any quota permission check.

Change-Id: I421ac8019cf16cc6f2a18ce6b4bcbcbf83397225
Bug-Url: https://bugzilla.redhat.com/1020201
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>