webadmin: Enable vNIC profile fields independently of version
Always enable the port mirroring and custom properties fields of the
vNIC profile dialog, independently of the DC compatibility
version. The vNIC profile is a DC-level entity and there could be
clusters inside the DC that do have a high enough compatibility
version, so the version compatibility should not be enforced here.

Change-Id: Id3f942133f752f91a601b305191749832efcf6f6
Bug-Url: https://bugzilla.redhat.com/1023952
Signed-off-by: Lior Vernia <lvernia@redhat.com>