webadmin: Small refactoring to simplify VnicProfileModel
It wasn't really clear what exactly dcId and dcCompatibilityVersion
were affecting, whereas customPropertiesVisible was passed around as a
parameter when it really should only be set when the VnicProfileModel
is constructed.

Also theoretically improved the behavior of selecting a QoS
configuration from the list. Currently when editing a network the DC
can't be changed, but if that is enabled in the future, then if the DC
will be switched and then switched back, the original QoS selection
will be salvaged.

Change-Id: I8e98a1fc26ce6b32076e09a67e6a3d82f52964df
Signed-off-by: Lior Vernia <lvernia@redhat.com>