userportal, webadmin: workaround for Firefox favicon bug
Workaround for Firefox bug https://bugzilla.mozilla.org/show_bug.cgi?id=519028,
'change location.hash leads to location bar icon (favicon) disappearance.'
The symptom is that the favicon disappears in Firefox on the initial hashchange
when logging in.

The workaround is to simply detach and re-attach the favicon link element
on hashchange by overriding ApplicationPlaceManager.onValueChange, and only
in Firefox by checking the user-agent.

Change-Id: Ia4b0d8a0176eb9299aec01ae6772297409ef3108
Bug-Url: https://bugzilla.redhat.com/999436
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>