core: snapshot disks should be part of the vm disk map only.
*Currently the vm populated DiskList should contain only it's actual
image disks.
The vm attached snapshot disks should be treated as LUN/shareable
disks.

Regardless to this change, it needs to be inspected if we want to
perform the filtering of the disks when loading it from the db or by
using ImagesHandler.filterImageDisks, as currently in many flows the
filtering is done although VM.getDiskList() contains only the needed
disks - therefore in this change I fixed only this part.

*When filtering the vm disks/devices - LUN disks should be filtered out
as well.

Change-Id: I5d18e5d3e321b871f4c7c283c8950fa7645eed5c
Related to Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=922475
Signed-off-by: Liron Aravot <laravot@redhat.com>