core: LSM - lock changes, added support for disk snapshot scenarios
This patch contains the following changes:
*When live migrating a disk, the disk (and it's snapshots) would be
locked for the whole flow preventing other operations to be done on the
disk. Therefore the create snapshot command will ignore those disks in
its checks and they won't be unlocked as the snapshot phase ends.

*Disk can be live migrated when snapshots of it are attached to other
vms unless those vms are not in "Down" status as it means that the disk
is being used by them.

*When live migrating the disk (after creating the snapshot for the
migration) all the disk snapshots would be locked and not just the
active volume.

*Task Handlers should be able to use the compensation context of the
enclosing command.

Change-Id: I0c7cb49ff37f28d0b3ab0df9f4ee1f3082aba018
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1028768
Signed-off-by: Liron Aravot <laravot@redhat.com>