webadmin: cannot import VM with snapshots without collapsing them
When importing a VM each disk of the VM is verified if it's parent is present
on a template in the system. If not, the snapshots has to be collapsed.

The problem was that if the VM had snapshots the disks returned in
getDiskMap() returned the snapshots (instead of the root disks).

Fixed by finding the root disks and the verification is done according to them.

Change-Id: Ib177757b64090bc5066acc3d89766bb0dbc29385
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1020830
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>