core: fix log message about rerun a vm
We used to have log message saying "vm {0} running in db and not running
in vds - add to rerun treatment" for each vm that was set as running in the
engine and reported as one that is not running by vdsm, even in cases where
we didn't add it to the 'rerun treatment' later on.

This patch change this logging message - now we log a message that just
say that the vm is running in db and not in vds in that case, and we log
another message saying that the vm is added to rerun treatment when it
is actually added to the 'rerun treatment'.

Change-Id: Icc2940845d150fcd41e7e327769f50c8bfffce5e
Signed-off-by: Arik Hadas <ahadas@redhat.com>