jsonrpc: Stomp changes in vdsbroker
Stomp reactor used in vdsbroker with protocol configuration stored in
new column in db.

There is ability to choose protocol from the UI or rest api when a new
host is added to the system. Chosen protocol is represented by 0 for xml
and 1 for stomp, There is plan to add amqp which probably will be 2.
Protocol values are represented as enum.

Bug-Url: https://bugzilla.redhat.com/1081049
Change-Id: If78de6620ba6891543531ac8ddd633b67828a89c
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>