core: change inner representation for tunnel migration
Findbugs complained about setting tunnel-migration to null, when it was
represented as Boolean. We used null value to indicate that the
'VdsProperties.TUNNELED' attribute should not be sent to VDSM (in order
to maintain backward compatibility).

This patch changes the representation of tunnel-migration from Boolean
to boolean, and in MigrateBrokerVDSCommand we check again if the
tunnel-migration feature is supported or not, and if it is not supported
then we don't send the 'VdsProperties.TUNNELED' to VDSM.

Change-Id: I9f1df36ec741f09f49a436854486d37c442fc042
Signed-off-by: Arik Hadas <ahadas@redhat.com>