core: change required permissions for removing VM from export domain
This patch changes the permissions which are required for removing a VM
from export domain. Until now, the user had to have import-export
permissions on the VM in order to be able to remove it from the export
domain - It was problematic in case the VM which is going to be removed
doesn't exist in the system.

From now on, we require 'delete vm' permissions on the export domain
in order to be able to remove VMs from the export domain, which makes it
possible for everyone who has 'delete vm' permissions on the export
domain to remove those VMs, also in case they don't exist in the DB.

Change-Id: I9f63cc6e53cd082d011bbef72c6a9435b6c1620b
Bug-Url: https://bugzilla.redhat.com/1010963
Signed-off-by: Arik Hadas <ahadas@redhat.com>