userportal: error flow of RDP RdpPluginImpl broken
There were the following issues which caused the error not to be propagated:
- a regression caused by 625b7452d840793df5e72764193c98c5ba121cdf
  which did bring back the crashing of IE if the RDP connection is not
  successful (originally fixed by fa10641c3484ea10d0b1b33ef8e3d656b5e6dd35).

- by the same commit (625b7452d840793df5e72764193c98c5ba121cdf) the new method
  setParentModel has been introduced but it was not called. It caused a NPE on
  RdpPluginImpl.rdpDisconnected:66. Fixed by calling this method on the
  relevant places.

This patch differs from the 34d4bbcbe8f01297b8770bf69b445b08dcf1b1f7
from which is this backported in a way that the AbstractRdp.getUserPassword
already handles the null correctly in this branch (e.g. no change required
in this class).

Change-Id: I85fc3b32c4f544f23864db2aefa5ebcd4d896b75
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1014710
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>