webadmin: Small fix to VnicProfilesEditor
As all the changes in the profiles are committed to the model as soon
as they happen (adding/removing rows, typing characters) it's
unnecessary to flush the widget; it also causes a ghost row to be
added the moment "OK" is pressed, so it would be better to avoid this
flicker.

Also added some javadoc to prevent usage of a method.

Change-Id: I504ebef37ef6e3d259cdea8d766a64b4aec51930
Signed-off-by: Lior Vernia <lvernia@redhat.com>