core,frontend: show template name also for cloned VMs
If the template provisioning of the VM is clone, the general subtab
of the VM/Pool shows that the "Template"'s value is Blank.
This patch changes this to the following behavior:
- for cloned provisioning the "Template" field shows:
  "templateName (Clone/Independent)"
- for thin provisioning the "Template" field shows:
  "templateName (Thin/Dependent)"
- for VMs created from Blank template the "Template" field shows:
  "templateName"
- for VMs created before this feature added the "Template" field shows:
  "templateName"
Same logic apply for the Storage main tab -> VM Import subtab ->
Template column
The implementation:
Two new fields added to the vm_static table:
- original_template_id: never changes
- original_template_name: changes when the template on which this VM
  is based on changes the name. This is checked during VM/template import
  and during template update
The original_template_name field is there and maintained "by hand" (instead of
having a reference to the template in original_template_id) is because this
field has to be shown even after the original template has been deleted.

Change-Id: Ieea4ad7c1ed845d3ff57aaafc19f0830ab72b2d1
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1042784
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>