webadmin: Dont replace . with _ if error not in resources
Modified to not to replace . with _ if the error message is NOT found in the
resources.
This was causing an issue if an error does not exist in resources, and
contains full stops or valid "." (e.g. IP Address / FQDN). All of the
"." were getting replaced with _ if the error message shown on the
screen.

Change-Id: I68c8354c7fec9e8a85fbea47bd7e4586f49defdf
Bug-Url: https://bugzilla.redhat.com/1024184
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>