webadmin: fix feedback image visiblity issue
'Feedback' section in the right-top will be hidden
if there is no feedback-url is available.
But that section is a collection images and text,
not all of them were hidden when no feedback-url supplied.

This is fixed by making the whole feedback panel
invisible if there is no feedback-url available.

Change-Id: I08c673909ba2a1c13e6bcac70a3ff19a99a1e6a3
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=968511
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>