webadmin: Changed by reference check to by value (findbugs)
When setting a model's change prohibition reason, the check whether
the value is different than the preceding value, in order to see if an
event has to be raised, was being done by reference rather than by
value. It doesn't really matter because all strings passed to the
method are constants, but the norm with triggering value change events
is to check by value.

Change-Id: Ibca529d23718dfaa5bf120d0bc157c259f016911
Signed-off-by: Lior Vernia <lvernia@redhat.com>