packaging: setup: preserve shmmax config on cleanup
Previously if you ran engine-setup; engine-cleanup; engine-setup
the final engine-setup may not properly store a persistent
configuration for shmmax in sysctl, so that a subsequent
reboot will cause postgresql to fail to come up so oVirt
is non-operational.

Now cleanup won't remove the configuration for shmmax in sysctl
just telling the user where it's placed and how to restore
default value manually. This will ensure that a working postgresql
will work also after reboot.

Change-Id: I4a95d5d9c678da80ee3f1e0a9de511309dfd4b51
Bug-Url: https://bugzilla.redhat.com/1026926
Bug-Url: https://bugzilla.redhat.com/1027605
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>