core: fix stateless VM which was reverted to ram snapshot
We were removing the memory volumes from snapshots in case the live part
of the snapshot did not happen and the snapshot is set to include
memory. This logic is not good because when it comes to stateless
snapshots, the snapshots contain memory (the memory that was associated
with the active snapshot) and it is not a live snapshot, but we don't
want to remove the memory associated with it.

Therefore, this patch changes the logic that is used to determine
whether or not to remove the memory which is associated with snapshot
when the live part of the snapshot doesn't happen, such that we'll also
ask if the memory was created as part of this snapshot and only if it
is, we will delete it.

Change-Id: I0a4e479424757fab1107797568bc019681de8265
Bug-Url: https://bugzilla.redhat.com/1004184
Signed-off-by: Arik Hadas <ahadas@redhat.com>