core: Set min connection pool size to 1
In a previous patch we changed the default minimum database connection
pool size to 75 (it used to be 1). This was done in order to have
database connections ready from the very beginning and thus improve
performance. But in practice this means that 75 PostgreSQL processes are
created wether they are needed or not. These consume approx 2.1 GiB of
RAM. In machines with a total of 4 GiB this evantually triggers the out
of memory killer of the kernel, which will in turn select the process
that is using more memory, which usually happens to be the engine
itself, so it is killed. This patch changes the default minimum size
back to 1 to avoid this issue.

Change-Id: I9de2c9e6999cb22a4be99eabb66a7216e673d0d2
Bug-Url: https://bugzilla.redhat.com/1026100
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>