core: move the validation of active iso domain on run vm
The validation which checks that there is an active ISO domain when the
VM is configured with 'iso path' was located in RunVmCommand, while it
should be located in RunVmValidator. Besides moving that validation to
the RunVmValidator class, a minor improvement is introduced: instead of
searching the id of the active ISO domain each time we need it (and
we're planning to use it more soon), it is now cached for the lifetime
of the RunVm command.

Change-Id: Ic1bc81686be01088aa2dde5c61e4b83c96643f73
Signed-off-by: Arik Hadas <ahadas@redhat.com>