core: DetachDiskFromVm - redundant checks and ctor usage
When executing DetachDiskFromVm command, the used parameters c'tor was
always called with true for "isPlugUnplug" property.
That means that the relevant checks within the command where always
relevant to that value only.

This patch removes the unneeded CDA check and move the callers of the
command to use the two parameters c'tor.

Change-Id: I48536a5baf4851c68f6d7053e4751d53a54f56a1
Signed-off-by: Liron Aravot <laravot@redhat.com>