webadmin: Refactoring in ProfilesInstanceTypeEditor
Replaced the non-standard edit() method with the standard one, that
had been previously marked as deprecated. This required introducing a
new class VnicInstancesModel, but it's better practice (now developers
don't have to be trusted to use the "correct" edit() method). Also
took the chance to parameterize all the related ListModel instances,
as the class is now generic.

Change-Id: I5951932037850a56f861bfced9199640153db785
Signed-off-by: Lior Vernia <lvernia@redhat.com>