webadmin: Refactoring in VnicProfilesEditor
Replaced the non-standard edit() method with the standard one, that
had been previously marked as deprecated. This required introducing a
new class NetworkProfilesModel, but it's better practice (now
developers don't have to be trusted to use the "correct" edit()
method). Also took the chance to parameterize all the related
ListModel instances, as the class is now generic.

Change-Id: Id8b101f869d90d5e10636ac70e653376dc29f72d
Signed-off-by: Lior Vernia <lvernia@redhat.com>