core: don't send cloud-init drive automatically
Currently cloud init is sent automatically on first run,
just as sysprep is sent for windows vms.

The problem is that currently cloud init configuration is not stored in
the engine db, while sysprep does (in sysprep files).

Until this gap is implemented, this patch changes the logic not to send
cloud init drive automatically, only on user request.

Change-Id: I94e36467b9f9002105e3eedda34c53ec43f590e1
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1014194
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>