core: Host shouldn't move to UP if it's already connected to other pool
When host fails to connect to the pool, it shouldn't move to UP if it's
already connected to other pool regardless of the master domain status.

Regardless of this change, the error handling in
ConnectStoragePoolVdsCommand should be changed as well as the error
handling in it's callers, this change handles InitVdsOnUp.

Change-Id: Ifd3d8268516b010f7831b59bbc6331cfd70efec1
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1023732
Signed-off-by: Liron Aravot <laravot@redhat.com>