core: Version.Op[In]Equality in UpdateStoragePool
Replaced calls to Version.OpEquality and Version.OpInequality with
version.equals(version2), as it's the standard Java way of doing
things, and is easier to read.

This can be done since oldSpmVersion is known to be not null be the
following logic:
1. It's retrieved from _oldStoragePool, which is assumed to be
   not-null under the CURRENT implementation.
2. This member is retrieved from the database, where the storage_pool
   table has a NOT NULL constraint on the compatibility_version
    column.

note that this also  has a slight theoretical performance benefit, as
nullability is no longer checked on the first operand, but it's
doubtful that this will be be noticeable.

Change-Id: Id36fc037c3597bb01c041850c4eb231a366f79da
Signed-off-by: Allon Mureinik <amureini@redhat.com>