core: Don't use Version.OpEquality in RunVmCommand
In all the (three) places Version.OpEquality is used in RunVmCommand, at
least one of the Version's can be deduced to be non-null.
Under these circumstances, using version.equals(version2) is better, as
it's the standard Java way of doing things, and is easier to read.

Also, note it has a slight theoretical performance benefit, as
nullability is no longer checked on the first operand, but it's doubtful
that this will be be noticeable.

Change-Id: I1e77770ace7f0e6fd06dcea502c427cb202c0f66
Signed-off-by: Allon Mureinik <amureini@redhat.com>