webadmin: rm Version.OpEquality from DC model
Replaced a call to Version.OpEquality with version.equals(version2),
as it's the standard Java way of doing things, and is easier to read.

This can be done since all the versions in getVersion().getItems()
are ultimately retrieved from
GetAvailableClusterVersionsByStoragePoolQuery, which is guaranteed to
return a List<Version> of non-null elements.

note that this also  has a slight theoretical performance benefit, as
nullability is no longer checked on the first operand, but it's
doubtful that this will be be noticeable.

Change-Id: I9a02b827e4c2d1165cb7bdc69cb681df323d4c6f
Signed-off-by: Allon Mureinik <amureini@redhat.com>