core: optimize setVdsGroupCompatibilityVersion
VDS and VM methods setVdsGroupCompatibilityVersion(Version) contain a
false optimization that only updates the member if the value were to
change.

Since assignment in Java is much cheaper than equality checking, simply
updating the value of the member, regardless of whether it changed or
not, would be much faster than this so-called optimization.

Change-Id: I0166fccf2188360b2e8812b462eb5c79d25aebe4
Signed-off-by: Allon Mureinik <amureini@redhat.com>