webadmin: Use /api/ instead of /api
The GUI uses the /api URL to get a RESTAPI session that can then be used
by plugins. When using basic authentication this means that the browser
will remember the credentials and send them back to any request for URLs
with paths starting with /, all. With the trailing slash it will send
them only to URLs ending with /api/.

Change-Id: I096a8769ff018f60366ab4ec8d2425c7f8d0f7e2
Bug-Url: https://bugzilla.redhat.com/1032508
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>