core: fix possible NPE on import VM with ram snapshot
In case the memory volumes are set to be imported to storage domain that
no disk was set to be imported to, we were getting NPE in
ImportVmCommand#setDomainsForMemoryImages since we assumed that the
storage domain exists in domain2requiredSize map but it doesn't exists
there in that case.

So from now on we will check whether the storage domain that the memory
volumes are going to be imported to exists in domain2requiredSize or
not and act accordingly in ImportVmCommand#setDomainsForMemoryImages.

Change-Id: I2a574266bdf6f06b7906c1338428071094430a9b
Bug-Url: https://bugzilla.redhat.com/1031371
Signed-off-by: Arik Hadas <ahadas@redhat.com>