core: Log filtering progress in scheduling
This logs the filtering step for each Host removal during the
scheduling operation to make debugging easier.

It also improves the error popup when scheduling
fails. The goal is to tell the user why his hosts
were removed from the decision process.

A logging is also added for the Pin To Host policy
unit.

https://bugzilla.redhat.com/show_bug.cgi?id=1002005

Change-Id: I4edcc33b0d280ab92e3c98d8d5a7a75386592e32
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1002005
Signed-off-by: Martin Sivak <msivak@redhat.com>
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>