frontend: override SPICE proxy on cluster and pool level
Until now, the only place where the SPICE proxy could be set was the global
configuration (SpiceProxyDefault).

This patch adds the possibility to override this option on cluster level (all
VMs and pools from this cluster will use it) and on pool level (all VMs from
the pool will use it).

Change-Id: I4d547a12b88e1acc8806fcab1f626b419a979742
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=974076
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>