core: Remove unused XmlDocument.nameTable
XmlDocument.nameTable is never initialized and therefore is always null.
Since it holds no data, it can safely be removed.

The only place this field was ever referenced was by calls to new
XmlNamespaceManager(Object), which simply ignored the passed object.
Since this parameter is not used anyway, and is always called with
XmlDocument.nameTable, the constructor was changed and the parameter was
removed.

Change-Id: Ie922fdd01d85a6f61fadaccd2a20263c7d013f85
Signed-off-by: Allon Mureinik <amureini@redhat.com>