core: SnapshotManager - disk devices might not represent engine disks
When using the snapshot manager there is an attempt to remove the disks
that aren't part of the snaposhot.

That's being done by loading the vm disk devices, load the disk for each
device and remove it if needed.

If the device represented disk isn't a disk known by the engine, do not
remove it (as being done for disks that aren't snappable/snapshot
disks).

Note - This method could be refactored to load the vm disks
directly instead of the devices and then the disks. In this patch
the null check is be added, while further refactor would be done in
follow up changes (the device are synced multiple times, the vm disks
could be loaded instead of the devices..etc).

Change-Id: Ie38c39c0f0409988a73610fe1dfb9a22afee5553
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1032807
Signed-off-by: Liron Aravot <laravot@redhat.com>