core,webadmin: add isVm and isTemplate methods to VmEntityType
Since now there are more types for VmEntityType (like instance type),
code that check for these types need to be updated.

Since the new entities are just types of templates,
and looking fwd we might add more like these,
looks like caller only want to know if this is a vm or a template.

So I added methods for easier reference,
and changed all references to specific types to the right method.

Change-Id: I6fc92031ab0afd08ab309a8ef6fd71a3fddb2204
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>