core: fix audit log on migration rerun failure case
In a specific case where migration fails, thus the VM is switched back
to Up state in the source and then goes down in the source, we used to
print an audit log message saying that it will be rerun although we
won't try to rerun. So this patch fix that audit log to be the one that
says the migration process failed.

Change-Id: I8e608cbdac5d38f7c3468f57cc563b8a3715359e
Signed-off-by: Arik Hadas <ahadas@redhat.com>